backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai>=0.8.0
PyMySQL==1.1.0
cryptography==41.0.7
PyJWT==2.8.0
redis==5.0.1
celery==5.3.6
reportlab==4.0.8
pandas==2.1.4
textstat==0.7.3
spacy==3.7.2
nltk==3.8.1
python-magic
werkzeug==3.0.1
Flask-JWT-Extended==4.6.0
python-docx==1.1.0
requests==2.31.0
